Removed lib from .gitignore as it ignores action/libs directory, moved json_serial into lib
